Move `CRIClient` to another package, keeping the runtime-client clean
